# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /social:
    get:
      operationId: getSocial
      tags: [social]
      summary: Get the full social/gameplay aggregate for the authenticated user
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Social aggregate
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/SocialData" }
    post:
      operationId: postSocialFallback
      deprecated: true
      tags: [social]
      summary: Deprecated no-op (state changes use /social/* action endpoints)
      security:
        - sessionCookie: []
      responses:
        "404":
          description: Use the specific /social/{action} endpoints
  /social/reveal/tap:
    post:
      operationId: revealTap
      tags: [social]
      summary: Curtain tap — server spends LP, advances reveal progress
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelId]
              properties: { modelId: { type: string } }
      responses:
        "200":
          description: Updated aggregate
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/SocialData" } }
        "400":
          description: Not enough LP / unknown model
  /social/reveal/unlock:
    post:
      operationId: revealUnlock
      tags: [social]
      summary: Instant RM Coin unlock — server validates cost and balance
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelId]
              properties: { modelId: { type: string } }
      responses:
        "200":
          description: Updated aggregate
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/SocialData" } }
        "400":
          description: Not enough RM / already unlocked
  /social/follow:
    post:
      operationId: followModel
      tags: [social]
      summary: Toggle a model follow
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelId]
              properties: { modelId: { type: string } }
      responses:
        "200":
          description: Updated aggregate
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/SocialData" } }
  /social/post:
    post:
      operationId: postInteraction
      tags: [social]
      summary: Toggle like/save/hot on a post
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [postId, flag, value]
              properties:
                postId: { type: string }
                flag: { type: string, enum: [liked, saved, hot] }
                value: { type: boolean }
      responses:
        "200":
          description: Updated aggregate
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/SocialData" } }
  /social/view:
    post:
      operationId: recordView
      tags: [social]
      summary: Increment a post view count (atomic)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [postId]
              properties: { postId: { type: string } }
      responses:
        "200":
          description: Updated aggregate
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/SocialData" } }
  /social/follow-style:
    post:
      operationId: followStyle
      tags: [social]
      summary: Toggle a style follow
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [styleId]
              properties: { styleId: { type: string } }
      responses:
        "200":
          description: Updated aggregate
  /social/mute:
    post:
      operationId: setMute
      tags: [social]
      summary: Toggle global video mute (non-authoritative metadata)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties: { value: { type: boolean } }
      responses:
        "200":
          description: Updated aggregate
  /social/wheel/spin:
    post:
      operationId: wheelSpin
      tags: [social]
      summary: Prize-wheel spin — SERVER picks the prize, validates spin availability
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Prize + updated aggregate
          content:
            application/json:
              schema:
                type: object
                properties:
                  prize: { type: object }
                  data: { $ref: "#/components/schemas/SocialData" }
        "409":
          description: No spins available
  /social/ads/intent:
    post:
      operationId: adIntent
      tags: [social]
      summary: Register a watched ad -> advances the earned-spin counter
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Updated aggregate
  /social/purchase:
    post:
      operationId: purchaseItem
      tags: [social]
      summary: Buy a miner/skin — server validates cost against catalogue
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId]
              properties: { itemId: { type: string } }
      responses:
        "200":
          description: Updated aggregate
        "400":
          description: Not enough RM / already owned / unknown item
  /social/equip:
    post:
      operationId: equipItem
      tags: [social]
      summary: Equip an owned miner/skin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId]
              properties: { itemId: { type: string } }
      responses:
        "200":
          description: Updated aggregate
        "400":
          description: Item not owned / unknown
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rm_session
  schemas:
...
    SocialData:
      type: object
      description: Canonical social/gameplay aggregate owned by the backend.
      properties:
        starterIds: { type: array, items: { type: string } }
        unlockedIds: { type: array, items: { type: string } }
        revealProgress: { type: object, additionalProperties: { type: number } }
        lpBalance: { type: number }
        followingIds: { type: array, items: { type: string } }
        likedPostIds: { type: array, items: { type: string } }
        hotPostIds: { type: array, items: { type: string } }
        savedPostIds: { type: array, items: { type: string } }
        followedStyles: { type: array, items: { type: string } }
        muted: { type: boolean }
        viewCounts: { type: object, additionalProperties: { type: number } }
        ownedMiners: { type: array, items: { type: string } }
        ownedSkins: { type: array, items: { type: string } }
        equippedMiner: { type: string, nullable: true }
        equippedSkin: { type: string, nullable: true }
        modelTickets: { type: integer }
        shotTickets: { type: integer }
        bonusMineMinutes: { type: integer }
        lastWheelSpin: { type: string, format: date-time, nullable: true }
        earnedSpins: { type: integer }
        adsWatchedForSpin: { type: integer }
...